Back features - refactoring and bugs

+++ Personal Blog/Controllers/ArticleController.cs
@@ -1,7 +1,8 @@
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
-using Personal_Blog.Model.Domain;
+using Personal_Blog.Model.Exceptions;
 using Personal_Blog.Model.Requests;
+using Personal_Blog.Model.Responses;
 using Personal_Blog.Services;
 
 namespace Personal_Blog.Controllers;
@@ -13,45 +14,75 @@ public class ArticleController(ILogger<ArticleController> logger, ArticleService
     private readonly ArticleService _articleService = articleService;
 
     [HttpGet("getAll")]
-    public async Task<ActionResult<IEnumerable<Article>>> GetAll()
+    public async Task<ActionResult<IEnumerable<ArticleResponse>>> GetAll()
     {
         var result = await _articleService.GetAll();
         return Ok(result);
     }
 
     [HttpGet("get/{id}")]
-    public async Task<ActionResult<Article>> GetById(string id)
+    public async Task<ActionResult<ArticleResponse>> GetById(string id)
     {
-        var result = await _articleService.GetById(id);
-        if (result == null)
+        if (string.IsNullOrEmpty(id) || id.Length != _articleService.GetRequiredIdLength())
         {
-            return NotFound(new{ Id = id });
+            return BadRequest();
         }
-        return Ok(result);
+        try
+        {
+            var result = await _articleService.GetById(id);
+
+            return Ok(result);
+        }
+        catch (ArticleNotfoundException ex)
+        {
+            return NotFound(new { message = ex.Message });
+        }
     }
 
     [HttpPost("add")]
     [Authorize(Roles = "Admin")]
-    public async Task<ActionResult> AddArticle(CreateArticleRequest request)
+    public async Task<ActionResult<ArticleResponse>> AddArticle([FromBody] CreateArticleRequest request)
     {
-        await _articleService.InsertArticle(request);
-        return Ok();
+        var result = await _articleService.InsertArticle(request);
+        return Ok(result);
     }
 
     [HttpDelete("delete/{id}")]
     [Authorize(Roles = "Admin")]
-    public async Task<ActionResult<Article>> DeleteArticle(string id)
+    public async Task<ActionResult<ArticleResponse>> DeleteArticle(string id)
     {
-        var result = await _articleService.DeleteById(id);
-        if (result) return Ok(new { Id = id });
-        else return NotFound();
+        if (string.IsNullOrEmpty(id) || id.Length != _articleService.GetRequiredIdLength())
+        {
+            return BadRequest();
+        }
+        try
+        {
+            var result = await _articleService.DeleteById(id);
+
+            return Ok(result);
+        }
+        catch (ArticleNotfoundException ex)
+        {
+            return NotFound(new { message = ex.Message });
+        }
     }
     [HttpPatch("update/{id}")]
     [Authorize(Roles = "Admin")]
-    public async Task<ActionResult<Article>> UpdateArticle(string id, [FromBody] UpdateArticleRequest request)
+    public async Task<ActionResult<ArticleResponse>> UpdateArticle(string id, [FromBody] UpdateArticleRequest request)
     {
-        var result = await _articleService.UpdateOne(id, request);
-        if (result) return Ok(new { Id = id });
-        return NotFound();
+        if (string.IsNullOrEmpty(id) || id.Length != _articleService.GetRequiredIdLength())
+        {
+            return BadRequest();
+        }
+        try
+        {
+            var result = await _articleService.UpdateOne(id, request);
+
+            return Ok(result);
+        }
+        catch (ArticleNotfoundException ex)
+        {
+            return NotFound(new { message = ex.Message });
+        }
     }
 }
